HIVE 22205 : Upgrade zookeeper and curator versions
Other components like hadoop have switched to using new ZK versions. So these jars end up in classpath for hive services and could cause issues due to in-compatible curator versions that hive uses.

So it makes sense for hive to upgrade the ZK and curator versions to try to keep up.